@@ -6,8 +6,8 @@
   "dependencies": {
     "@reach/rect": "^0.16.0",
     "focus-visible": "^5.2.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-lazyload": "^3.2.0",
     "react-router-dom": "^5.2.0",
     "react-router-last-location": "^2.0.1",
